raspberry_pi/requirements.txt
paho-mqtt>=1.6.1
pyyaml>=6.0
scikit-learn>=1.3.0
requests>=2.31.0
